Use ExponentialRetry
