+++ src/Telemetry/AzureDevOpsClient/AzureDevOpsClientFactory.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using Microsoft.Extensions.Logging;
 using System.Threading.Tasks;
 
 namespace Microsoft.DotNet.Internal.AzureDevOps
@@ -12,7 +13,7 @@ namespace Microsoft.DotNet.Internal.AzureDevOps
 
         public IAzureDevOpsClient CreateAzureDevOpsClient(string baseUrl, string organization, int maxParallelRequests, string accessToken)
         {
-            return new AzureDevOpsClient(baseUrl, organization, maxParallelRequests, accessToken);
+            return new AzureDevOpsClient(baseUrl, organization, maxParallelRequests, accessToken, null);
         }
     }
 }
