+++ src/Telemetry/AzureDevOpsClient/IAzureDevOpsClient.cs
@@ -5,6 +5,7 @@
 using System;
 using System.IO;
 using System.Net.Http;
+using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -20,6 +21,6 @@ namespace Microsoft.DotNet.Internal.AzureDevOps
         public Task<Timeline> GetTimelineAsync(string project, int buildId, string timelineId, CancellationToken cancellationToken);
         public Task<BuildChangeDetail> GetChangeDetails(string changeUrl, CancellationToken cancellationToken = default);
         public Task<WorkItem> CreateRcaWorkItem(string project, string title, CancellationToken cancellationToken = default);
-        public Task<HttpResponseMessage> TryGetLogContents(string logUri, CancellationToken cancellationToken = default);
+        public Task<string> TryGetImageName(string logUri, Regex imageNameRegex, CancellationToken cancellationToken = default);
     }
 }
